# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Torneo Solidario",
        "description": "API del torneo solidario",
        "contact": {
            "email": "minillanillo@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/jugadores": {
            "get": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Obtener todos los jugadores de la web",
                "description": "Obtener todos los jugadores en la llamada a la API",
                "operationId": "indexJugadores",
                "responses": {
                    "200": {
                        "description": "Jugadores disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugadores disponibles"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Jugador"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay jugadores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No hay jugadores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Crear un nuevo jugador",
                "description": "Crear un nuevo jugador",
                "operationId": "storeJugador",
                "requestBody": {
                    "description": "Datos del jugador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Jugador 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "jugador creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador creado correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Jugador"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jugadores/{id}": {
            "get": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Obtener un jugador",
                "description": "Obtener un jugador por su id",
                "operationId": "showJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "jugador encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador encontrado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Jugador"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jugador no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Actualizar un jugador",
                "description": "Actualizar un jugador",
                "operationId": "updateJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del jugador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Jugador 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jugador actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador actualizado correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Jugador"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jugador no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Eliminar un jugador",
                "description": "Eliminar un jugador por su id",
                "operationId": "deleteJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jugador no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Jugador no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ciclo": {
                "title": "Ciclo",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "FPGB Tapicería y Cortinaje"
                    },
                    "familia": {
                        "$ref": "#/components/schemas/Familia"
                    }
                },
                "type": "object"
            },
...
            "Estudio": {
                "title": "Estudio",
                "properties": {
                    "centro": {
                        "type": "string",
                        "example": "IES Ejemplo"
                    },
                    "curso": {
                        "type": "integer",
                        "example": 1
                    },
                    "ciclos": {
                        "$ref": "#/components/schemas/Ciclo"
                    }
                },
                "type": "object"
            },
            "Familia": {
                "title": "Familia",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Textil, Confección y Piel"
                    }
                },
                "type": "object"
            },
            "Jugador": {
                "title": "Jugador",
                "required": [
                    "nombre",
                    "equipo_id"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Álvaro"
                    },
                    "apellido1": {
                        "type": "string",
                        "example": "Ruiz"
                    },
                    "apellido2": {
                        "type": "string",
                        "example": "Gutierrez"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "jugador/capitan/entrenador"
                    },
                    "dni": {
                        "type": "string",
                        "example": "12345678A"
                    },
                    "email": {
                        "type": "string",
                        "example": "example@exa.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+34 666 666 666"
                    },
                    "goles": {
                        "type": "intenger",
                        "example": 1
                    },
                    "asistencias": {
                        "type": "integer",
                        "example": 1
                    },
                    "tarjetas_amarillas": {
                        "type": "integer",
                        "example": 1
                    },
                    "tarjetas_rojas": {
                        "type": "integer",
                        "example": 1
                    },
                    "lesiones": {
                        "type": "integer",
                        "example": 1
                    },
                    "equipo_nombre": {
                        "type": "string",
                        "example": "Equipo Example"
                    },
                    "estudios": {
                        "$ref": "#/components/schemas/Estudio"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to authenticate",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}